+++ api/hooks/registerSerializers.js
@@ -1,0 +1,4 @@
+import { Hooks } from './../../index';
+const { registerSerializers } = Hooks;
+
+export default registerSerializers;