+++ templates/actions/destroy.js
@@ -15,6 +15,11 @@ module.exports = function(interrupts = {}) {
   interrupts.destroy = interrupts.destroy ? interrupts.destroy : defaultInterrupt;
 
   return function(req, res) {
+    // Set the JSONAPI required header
+    // Technically because there is no content, no content-type header is required and in fact is being
+    // actively removed when calling res.send
+    // res.set('Content-Type', 'application/vnd.api+json');
+
     const Model = actionUtil.parseModel(req);
     const pk = actionUtil.requirePk(req);
     const query = Model.findOne(pk);
@@ -40,7 +45,7 @@ module.exports = function(interrupts = {}) {
             Model._retire(record);
           }
         }
-        interrupts.destroy.call(this, req, res, () => res.ok(null, actionUtil.parseLocals(req)), Model, record);
+        interrupts.destroy.call(this, req, res, () => res.noContent(null, actionUtil.parseLocals(req)), Model, record);
         // @todo --- if neccessary, destroy related records
       });
     });