+++ templates/actions/find.js
@@ -17,6 +17,9 @@ module.exports = function(interrupts = {}) {
   interrupts.find = interrupts.find ? interrupts.find : defaultInterrupt;
 
   return function(req, res) {
+    // Set the JSONAPI required header
+    res.set('Content-Type', 'application/vnd.api+json');
+
     // Look up the model
     const Model = actionUtil.parseModel(req);
     // parse criteria from request
@@ -73,10 +76,7 @@ module.exports = function(interrupts = {}) {
                   actionUtil.subscribeDeep(req, record);
                 });
               }
-              const emberizedJSON = Ember.buildResponse(Model, matchingRecords, associations, associated);
-              emberizedJSON.meta = {
-                total: results.count
-              };
+              const emberizedJSON = Ember.buildResponse(Model, matchingRecords, { total: results.count });
               res.ok(emberizedJSON, actionUtil.parseLocals(req));
             },
             Model,