+++ templates/actions/findone.js
@@ -16,6 +16,9 @@ module.exports = function(interrupts = {}) {
   interrupts.findone = interrupts.findone ? interrupts.findone : defaultInterrupt;
 
   return function(req, res) {
+    // Set the JSONAPI required header
+    res.set('Content-Type', 'application/vnd.api+json');
+
     const Model = actionUtil.parseModel(req);
     const pk = actionUtil.requirePk(req);
     const query = Model.findOne(pk);
@@ -52,7 +55,7 @@ module.exports = function(interrupts = {}) {
               Model.subscribe(req, [matchingRecord[Model.primaryKey]]);
               actionUtil.subscribeDeep(req, matchingRecord);
             }
-            res.ok(Ember.buildResponse(Model, matchingRecord, associations, associated), actionUtil.parseLocals(req));
+            res.ok(Ember.buildResponse(Model, matchingRecord), actionUtil.parseLocals(req));
           },
           Model,
           matchingRecord