+++ test/fixtures/Article.js
@@ -2,7 +2,7 @@ export default [
   {
     title: 'JSON to XML',
     author: '1',
-    comments: ['1', '2']
+    comments: ['1', '2', '3']
   },
   {
     title: 'JSON API: Whaaa?',