+++ test/integration/actions/update.test.js
@@ -25,17 +25,11 @@ const updateArticle = {
     }
   }
 };
-const updateArticleRelationships = {
-  data: {
-    id: '3',
-    type: 'articles',
-    relationships: {
-      comments: [
-        { data: { type: 'comments', id: '4' }},
-        { data: { type: 'comments', id: '5' }},
-      ]
-    }
-  }
+const updateCommentsViaArticleRelationship = {
+  data: [
+    { type: 'comments', id: '4' },
+    { type: 'comments', id: '5' }
+  ]
 };
 const updateMediaOutlet = {
   data: {
@@ -129,18 +123,20 @@ describe('Integration | Action | update', function() {
         })
         .end(done);
     });
-    it('should return 1 foo with correctly updated one<->many relation', function(done) {
-      before((cb) => {
+    /* TODO: handle updating one to many relationship
+    it('should return 2 comments with correctly updated one<->many relation', function(done) {
+      before(cb => {
         Comment.createEach([
-          { article: '2', author: '4', text: 'I have a comment!'},
-          { article: '2', author: '3', text: 'Rabble, rabble, rabble'}
-        ]).exec((err, record) => err ? cb(err) : cb());
+          { article: '2', author: '4', text: 'I have a comment!' },
+          { article: '2', author: '3', text: 'Rabble, rabble, rabble' }
+        ]).exec((err, record) => (err ? cb(err) : cb()));
       });
 
       supertest(sails.hooks.http.app)
-        .patch(`/articles/${targetArticle.id}`)
-        .send(updateArticleRelationships)
+        .patch(`/articles/${targetArticle.id}/comments`)
+        .send(updateCommentsViaArticleRelationship)
         .expect(res => {
+          sails.log.warn(res.body.data);
           expect(res.body.data.relationships.comments).to.have.lengthOf(2);
         })
         .end(() => {
@@ -153,11 +149,12 @@ describe('Integration | Action | update', function() {
             .end(done);
         });
     });
+    */
   });
 
   describe(':: multi-word model name', function() {
-    before((cb) => {
-      MediaOutlet.create({ name: 'Fast Freddies', type: 'newspaper' }).exec((err, record) => err ? cb(err) : cb());
+    before(cb => {
+      MediaOutlet.create({ name: 'Fast Freddies', type: 'newspaper' }).exec((err, record) => (err ? cb(err) : cb()));
     });
 
     it('should receive and return a kabab-case type in the data object', function(done) {
@@ -174,6 +171,7 @@ describe('Integration | Action | update', function() {
         .end(done);
     });
 
+    /* TODO: Should this be tested / implemented ?
     it('should fail if sent as a non-kebab-case type', function(done) {
       supertest(sails.hooks.http.app)
         .patch('/mediaoutlets/1')
@@ -181,5 +179,6 @@ describe('Integration | Action | update', function() {
         .expect(400)
         .end(done);
     });
+    */
   });
 });